requirements.txt
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
groq>=0.4.0
python-dotenv>=1.0.0
numpy>=1.24.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

